Firewalld is functioning properly I believe
Have adjusted firewalld to support rich_rules done separately, adjusted zone purging to reload firewalld instead of restart service.
Rich rules diff is much more readable now as well.  Converted masquerade to be boolean, it was a single boolean value inside an array previously.  Types must also convert the boolean true or false to symbol :true or :false to work, done.
TODO. Clear out comments and any remaining debug puts
